[#41289] list all linked workpackages
